+++ packages/client/hooks/useTipTapCommentEditor.ts
@@ -8,6 +8,7 @@ import Atmosphere from '../Atmosphere'
 import {LoomExtension} from '../components/promptResponse/loomExtension'
 import {TiptapLinkExtension} from '../components/promptResponse/TiptapLinkExtension'
 import {mentionConfig} from '../shared/tiptap/serverTipTapExtensions'
+import {ClearOnSubmit} from '../tiptap/extensions/ClearOnSubmit'
 import {tiptapEmojiConfig} from '../utils/tiptapEmojiConfig'
 import {tiptapMentionConfig} from '../utils/tiptapMentionConfig'
 import {tiptapTagConfig} from '../utils/tiptapTagConfig'
@@ -54,6 +55,7 @@ export const useTipTapCommentEditor = (
         TiptapLinkExtension.configure({
           openOnClick: false
         }),
+        ClearOnSubmit,
         onEnter &&
           onEscape &&
           Extension.create({