+++ packages/server/email/MailManagerMailgun.ts
@@ -1,6 +1,6 @@
 import FormData from 'form-data'
 import Mailgun from 'mailgun.js'
-import sendToSentry from '../utils/sendToSentry'
+import logError from '../utils/logError'
 import MailManager, {MailManagerOptions} from './MailManager'
 
 export default class MailManagerMailgun extends MailManager {
@@ -25,7 +25,7 @@ export default class MailManagerMailgun extends MailManager {
       })
     } catch (e) {
       const error = e instanceof Error ? e : new Error('Mailgun failed to create message')
-      sendToSentry(error, {tags: {to: toStr, type: 'Mailgun error'}})
+      logError(error, {tags: {to: toStr, type: 'Mailgun error'}})
       return false
     }
     return true
@@ -37,7 +37,7 @@ export default class MailManagerMailgun extends MailManager {
       return ['deliverable', 'catch_all', 'unknown'].includes(res.result)
     } catch (e) {
       const error = e instanceof Error ? e : new Error('Mailgun failed to validate emails')
-      sendToSentry(error, {tags: {type: 'Mailgun error'}})
+      logError(error, {tags: {type: 'Mailgun error'}})
       return false
     }
   }