+++ packages/server/graphql/private/queries/suOrgCount.ts
@@ -8,9 +8,10 @@ const suOrgCount: QueryResolvers['suOrgCount'] = async (_source, {minOrgSize, ti
       qb
         .selectFrom('Organization as o')
         .innerJoin('OrganizationUser as ou', 'o.id', 'ou.orgId')
+        .innerJoin('User as u', 'ou.userId', 'u.id')
         .select(({fn}) => fn.count('ou.id').as('orgSize'))
         .where('tier', '=', tier)
-        .where('ou.inactive', '=', false)
+        .where('u.inactive', '=', false)
         .where('ou.removedAt', 'is', null)
         .groupBy('o.id')
         .having(({eb, fn}) => eb(fn.count('ou.id'), '>=', minOrgSize))