+++ packages/server/types/modules.d.ts
@@ -38,7 +38,6 @@ interface Window {
     atlassian: string
     github: string
     google: string
-    sentry: string
     slack: string
     oauth2Redirect: string
     stripe: string